app/ocr.py
import easyocr
import numpy as np
from PIL import Image
import io

reader = easyocr.Reader(['en', 'hi'], gpu=False)

def extract_text_from_image(image_bytes: bytes) -> str:
    image = Image.open(io.BytesIO(image_bytes))
    image_np = np.array(image)

    results = reader.readtext(image_np)
    return " ".join([text[1] for text in results])
