requirement.txt
fastapi
uvicorn
easyocr
google-generativeai
pillow
numpy
python-multipart
jinja2
python-dotenv